refactor(providers): split content-moderation providers out of app_providers.dart
Sixth batch of #4506. Moves the 11-entity content-moderation cluster to
`lib/providers/moderation_providers.dart`:

- contentPolicyEngine
- divineHostFilterServiceProvider, divineHostFilterVersionProvider
- ageVerificationService
- contentFilterService, contentFilterVersion
- accountLabelService
- moderationLabelService
- contentBlocklistRepository
- BlocklistVersion (notifier class)
- blocklistSyncBridge (@keepAlive eager — activated by app_shell)

The file-private `_createBlockedAuthorFilter` helper stays in
`app_providers.dart` because three remaining video-domain providers
there still call it. Same posture as `videoFilterBuilder` staying with
batch 5.

`app_providers.dart` adds both `import 'moderation_providers.dart';`
and `export 'moderation_providers.dart';`. The import is required
because `_createBlockedAuthorFilter` and at least five other internal
sites (videoEventService, videosRepository, videoEventPublisher, etc.)
still consume the moved provider symbols. The export keeps the 74
existing consumer files (38 lib + 36 test) working unchanged.

Removed now-unused imports from `app_providers.dart`:
content_blocklist_repository, account_label_service,
age_verification_service, content_filter_service,
moderation_label_service. Kept content_policy package import
(BlockedVideoFilter helper still uses ContentPolicyEngine type).

Codegen moved: 9 *Provider chunks relocated from `app_providers.g.dart`
into `moderation_providers.g.dart`.

Two follow-up findings (added to existing lists):
1. blocklistSyncBridge contains repository-layer state-machine logic
   embedded in a provider — same shape as the dirty-sync bridge (batch
   4) and relay bridges (batch 5). Candidates for extraction into
   coordinator services.
2. Three providers (age/content/account) use `service.initialize();`
   fire-and-forget without `unawaited()`; moderationLabelService uses
   `unawaited(service.initialize().then(...))`. Style inconsistency to
   normalize.

Bridge error handling is policy-compliant: `Log.warning` for the
expected network/IO sync failure (per error_handling.md matrix).

Refs #4506, #4339 Wave 2
