+++ mobile/lib/providers/app_providers.dart
@@ -9,7 +9,6 @@ import 'package:blossom_upload_service/blossom_upload_service.dart';
 import 'package:categories_repository/categories_repository.dart';
 import 'package:collaborator_repository/collaborator_repository.dart';
 import 'package:comments_repository/comments_repository.dart';
-import 'package:content_blocklist_repository/content_blocklist_repository.dart';
 import 'package:content_policy/content_policy.dart';
 import 'package:curated_list_repository/curated_list_repository.dart';
 import 'package:curation_repository/curation_repository.dart';
@@ -38,6 +37,7 @@ import 'package:openvine/providers/app_foreground_provider.dart';
 import 'package:openvine/providers/curation_providers.dart';
 import 'package:openvine/providers/database_provider.dart';
 import 'package:openvine/providers/environment_provider.dart';
+import 'package:openvine/providers/moderation_providers.dart';
 import 'package:openvine/providers/nostr_client_provider.dart';
 import 'package:openvine/providers/og_viner_cache_provider.dart';
 import 'package:openvine/providers/preferences_providers.dart';
@@ -45,8 +45,6 @@ import 'package:openvine/providers/relay_providers.dart';
 import 'package:openvine/providers/saved_sounds_provider.dart';
 import 'package:openvine/providers/shared_preferences_provider.dart';
 import 'package:openvine/services/account_deletion_service.dart';
-import 'package:openvine/services/account_label_service.dart';
-import 'package:openvine/services/age_verification_service.dart';
 import 'package:openvine/services/analytics_service.dart';
 import 'package:openvine/services/api_service.dart';
 import 'package:openvine/services/auth_service.dart' hide UserProfile;
@@ -62,12 +60,10 @@ import 'package:openvine/services/collaborator_invite_service.dart';
 import 'package:openvine/services/collaborator_invite_state_store.dart';
 import 'package:openvine/services/collaborator_response_service.dart';
 import 'package:openvine/services/content_deletion_service.dart';
-import 'package:openvine/services/content_filter_service.dart';
 import 'package:openvine/services/content_reporting_service.dart';
 import 'package:openvine/services/crash_reporting_service.dart';
 import 'package:openvine/services/crosspost_api_client.dart';
 import 'package:openvine/services/curated_list_service.dart';
-import 'package:openvine/services/divine_host_filter_service.dart';
 import 'package:openvine/services/draft_storage_service.dart';
 import 'package:openvine/services/email_verification_listener.dart';
 import 'package:openvine/services/event_router.dart';
@@ -76,7 +72,6 @@ import 'package:openvine/services/hashtag_service.dart';
 import 'package:openvine/services/immediate_completion_helper.dart';
 import 'package:openvine/services/media_auth_interceptor.dart';
 import 'package:openvine/services/media_viewer_auth_service.dart';
-import 'package:openvine/services/moderation_label_service.dart';
 import 'package:openvine/services/mute_service.dart';
 import 'package:openvine/services/nip98_auth_service.dart';
 import 'package:openvine/services/nostr_creator_binding_service.dart';
@@ -114,6 +109,7 @@ import 'package:sound_service/sound_service.dart';
 import 'package:unified_logger/unified_logger.dart';
 import 'package:videos_repository/videos_repository.dart';
 
+export 'moderation_providers.dart';
 export 'nostr_apps_providers.dart';
 export 'notifications_providers.dart';
 export 'permissions_providers.dart';
@@ -121,11 +117,6 @@ export 'preferences_providers.dart';
 export 'relay_providers.dart';
 
 part 'app_providers.g.dart';
-
-@Riverpod(keepAlive: true)
-ContentPolicyEngine contentPolicyEngine(Ref ref) {
-  return ContentPolicyEngine.defaultRules();
-}
 
 BlockedVideoFilter _createBlockedAuthorFilter(Ref ref) {
   final blocklistRepository = ref.watch(contentBlocklistRepositoryProvider);
@@ -334,120 +325,6 @@ AnalyticsService analyticsService(Ref ref) {
   return service;
 }
 
-/// Divine-hosted-only filter preference service.
-final divineHostFilterServiceProvider = Provider<DivineHostFilterService>((
-  ref,
-) {
-  final prefs = ref.watch(sharedPreferencesProvider);
-  final service = DivineHostFilterService(prefs);
-  ref.onDispose(service.dispose);
-  return service;
-});
-
-/// Rebuild trigger for consumers that need to react to Divine-host filter
-/// preference changes.
-final divineHostFilterVersionProvider = Provider<int>((ref) {
-  final service = ref.watch(divineHostFilterServiceProvider);
-  var version = 0;
-
-  void listener() {
-    version++;
-    ref.invalidateSelf();
-  }
-
-  service.addListener(listener);
-  ref.onDispose(() => service.removeListener(listener));
-  return version;
-});
-
-/// Age verification service for content creation restrictions
-/// keepAlive ensures the service persists and maintains in-memory verification state
-/// even when widgets that watch it dispose and rebuild
-@Riverpod(keepAlive: true)
-AgeVerificationService ageVerificationService(Ref ref) {
-  final service = AgeVerificationService();
-  service.initialize(); // Initialize asynchronously
-  return service;
-}
-
-/// Content filter service for per-category Show/Warn/Hide preferences.
-/// keepAlive ensures preferences persist and are consistent across the app.
-@Riverpod(keepAlive: true)
-ContentFilterService contentFilterService(Ref ref) {
-  final ageVerificationService = ref.watch(ageVerificationServiceProvider);
-  final service = ContentFilterService(
-    ageVerificationService: ageVerificationService,
-  );
-  service.initialize(); // Initialize asynchronously
-  ref.onDispose(service.dispose);
-  return service;
-}
-
-/// Tracks content filter preference changes. Feed providers watch this
-/// to rebuild when the user changes a Show/Warn/Hide setting.
-@Riverpod(keepAlive: true)
-int contentFilterVersion(Ref ref) {
-  final service = ref.watch(contentFilterServiceProvider);
-  final aspectRatioPreference = ref.watch(
-    feedAspectRatioPreferenceServiceProvider,
-  );
-  var version = 0;
-  void listener() {
-    version++;
-    ref.invalidateSelf();
-  }
-
-  service.addListener(listener);
-  aspectRatioPreference.addListener(listener);
-  ref.onDispose(() {
-    service.removeListener(listener);
-    aspectRatioPreference.removeListener(listener);
-  });
-  return version;
-}
-
-/// Account label service for self-labeling content (NIP-32 Kind 1985).
-@Riverpod(keepAlive: true)
-AccountLabelService accountLabelService(Ref ref) {
-  final authService = ref.watch(authServiceProvider);
-  final nostrClient = ref.watch(nostrServiceProvider);
-  final service = AccountLabelService(
-    authService: authService,
-    nostrClient: nostrClient,
-  );
-  service.initialize();
-  return service;
-}
-
-/// Moderation label service for subscribing to Kind 1985 labeler events.
-@Riverpod(keepAlive: true)
-ModerationLabelService moderationLabelService(Ref ref) {
-  final nostrClient = ref.watch(nostrServiceProvider);
-  final authService = ref.watch(authServiceProvider);
-  final prefs = ref.watch(sharedPreferencesProvider);
-  final followRepository = ref.watch(followRepositoryProvider);
-  final service = ModerationLabelService(
-    nostrClient: nostrClient,
-    authService: authService,
-    sharedPreferences: prefs,
-  );
-  unawaited(
-    service.initialize().then((_) {
-      return service.syncFollowedLabelers(followRepository.followingPubkeys);
-    }),
-  );
-  final followingSubscription = followRepository.followingStream.listen((
-    pubkeys,
-  ) {
-    unawaited(service.syncFollowedLabelers(pubkeys));
-  });
-  ref.onDispose(() {
-    followingSubscription.cancel();
-    service.dispose();
-  });
-  return service;
-}
-
 /// Secure key storage service (foundational service)
 @Riverpod(keepAlive: true)
 SecureKeyStorage secureKeyStorage(Ref ref) {
@@ -573,94 +450,6 @@ PersonalEventCacheService personalEventCacheService(Ref ref) {
 @riverpod
 SeenVideosService seenVideosService(Ref ref) {
   return SeenVideosService();
-}
-
-/// Content blocklist service for filtering unwanted content from feeds
-///
-/// Injects SharedPreferences for local block persistence across restarts.
-/// Nostr publishing (kind 30000) is initialized via [syncBlockListsInBackground]
-/// during app startup in main.dart.
-///
-/// keepAlive ensures the relay subscription created by
-/// [syncBlockListsInBackground] survives widget rebuilds. Without it the
-/// provider auto-disposes, the subscription is lost, and blocks restored
-/// from the relay are never delivered to new instances.
-@Riverpod(keepAlive: true)
-ContentBlocklistRepository contentBlocklistRepository(Ref ref) {
-  final prefs = ref.watch(sharedPreferencesProvider);
-  return ContentBlocklistRepository(
-    prefs: prefs,
-    onChanged: () {
-      if (!ref.mounted) return;
-      ref.read(blocklistVersionProvider.notifier).increment();
-    },
-  );
-}
-
-/// Version counter to trigger rebuilds when blocklist changes.
-/// Widgets watching this will rebuild when block/unblock actions occur.
-@riverpod
-class BlocklistVersion extends _$BlocklistVersion {
-  @override
-  int build() => 0;
-
-  void increment() => state++;
-}
-
-/// Bridge that starts blocklist sync when the Nostr session becomes ready.
-///
-/// Watch this at app shell level. It listens to [nostrSessionProvider] and
-/// triggers [syncMuteListsInBackground] + [syncBlockListsInBackground]
-/// the first time the signer-backed Nostr client is initialized. This covers:
-/// - Already-authenticated startup (iOS keychain persists across reinstalls)
-/// - Post-login authentication (Android wipes credentials on uninstall)
-///
-/// Both sync methods have internal guards (`_mutualMuteSyncStarted`,
-/// `_blockListSyncStarted`) so duplicate calls are no-ops.
-@Riverpod(keepAlive: true)
-void blocklistSyncBridge(Ref ref) {
-  final authService = ref.watch(authServiceProvider);
-  final blocklistRepository = ref.watch(contentBlocklistRepositoryProvider);
-
-  Future<void> startSync(NostrSessionReadiness readiness) async {
-    final pubkey = readiness.pubkey;
-    final client = readiness.client;
-    if (!readiness.isReadyForActiveClient || pubkey == null || client == null) {
-      return;
-    }
-
-    if (authService.currentIdentity?.pubkey != pubkey) {
-      return;
-    }
-
-    try {
-      await Future.wait([
-        blocklistRepository.syncMuteListsInBackground(client, pubkey),
-        blocklistRepository.syncBlockListsInBackground(
-          client,
-          authService,
-          pubkey,
-        ),
-      ]);
-      Log.info(
-        '[BRIDGE] Block/mute list sync started',
-        name: 'BlocklistSyncBridge',
-        category: LogCategory.system,
-      );
-    } catch (e) {
-      Log.warning(
-        '[BRIDGE] Block/mute list sync failed (non-critical): $e',
-        name: 'BlocklistSyncBridge',
-        category: LogCategory.system,
-      );
-    }
-  }
-
-  unawaited(startSync(ref.read(nostrSessionProvider)));
-
-  ref.listen<NostrSessionReadiness>(nostrSessionProvider, (_, next) {
-    unawaited(startSync(next));
-  });
 }
 
 /// Draft storage service for persisting vine drafts
